Only generate lenses for elp-supported files
Summary:
This is to prevent crashes we are seeing trying to generate lenses for file of kind `OutsideProjectModel`.
This diff filters to only try for files known to be supported by ELP.
We are not able to do this at a general request level, because we don't have a `FileId` there.

Discussion of alternatives welcome

Reviewed By: ir-regular

Differential Revision: D56634386

fbshipit-source-id: 1247e3419d56ad2dec1008e62e30a4fdce203db2