Fix branch change diagnostic problem
Summary:
When we switch branches, ELP notices if any project config files have changed and reloads the project information if so.  By definition, this information is likely to be different.  In particular, there may be a different number of applications.  So the mapping of `FileId` to `AppData`, which is indexed by the `SourceRoot` associated with the file, may change.  If so we can get spurious diagnostics reported, as we try to compile a file with the incorrect configuration, such as include paths.

Prevent this, by forcing a re-mapping of this information when we switch workspaces.

Reviewed By: michalmuskala

Differential Revision: D56753435

fbshipit-source-id: 20da8786d7a8776c7cf11448adfb6ba0fbbf6d3f