combine related errors for eqwalizer cli reporting too
Summary:
The `DiagnosticsCollection` has functionality to combine related errors, so that if a function has a syntax error, the cascade of undefined functions from that reported by the erlang service compiler is collapsed, as they get added as related diagnostics to the syntax error.

Route the erlang service diagnostics reported by eqwalizer cli through this too.

Reviewed By: robertoaloi

Differential Revision: D56818059

fbshipit-source-id: 674d2622a12abf4a2d52339f0ec7fe8bb3c68791