Generate dynamically build_info file
Summary:
Add three more options to the `.elp.toml` configuration:

* `generate_build_info`
* `build_info_apps`
* `build_info_deps`

For projects which do not use `rebar3` or `buck2` as a build system, it is a lot of work to create and maintain a `build_info.json` file manually.

While ideally one would have an actual integration with the custom build tool, these three arguments are usually enough to guarantee basic functionalities for a project, as proven by the classic `erlang_ls.config` configuration mechanism.

Reviewed By: alanz

Differential Revision: D57379738

fbshipit-source-id: d0f55b2efcd9fc45dd82e2e4a14f1d7f97faa3e2