Add config option to select buck project model query
Summary:
We currently load the buck project model into ELP using a hard-code buck2 uquery inside ELP.
This comes close to matching the environment for buck2 when building, but is not completely accurate.
In particular, it does not set all include paths correctly, resulting in periodic CI failures when `elp parse-all` fails due to the mismatch.

The fix is being applied in two stages
1. Migrate the base query into a buck BXL query
2. Extend the query to include the additional include information

Given the centrality of project model loading to the ELP experience, we are taking a slow approach to rolling it out.  The plan is to introduce two configuration settings, one to control whether to use a bxl query at all, or use the current mechanism, the other to select between the vanilla BXL query or the one with extra include information.

This diff does a lot of plumbing to ensure these two settings are available at the correct point in buck project loading, whether from a test fixture, CLI invocation or LSP server.

The nett result is to have no effect, as the defaults are to stick with the current usage.

The `todo!()` items will be resolved in a subsequent diff

Reviewed By: robertoaloi

Differential Revision: D57277985

fbshipit-source-id: 1872f4b5aa489947651fcd635f842a98cd54b3b5