Fuzzy (any arity) go-to-definition for functions
Summary: When a function definition is (or function references are) not found, fall back to the first function with the same name but different arity, instead of returning "no results". This should improve ergonomics significantly while authoring Erlang code and is in line with other IDE experiences such as Rust Analyzer.

Reviewed By: alanz

Differential Revision: D57864572

fbshipit-source-id: ae552f4d4f5ac142141eb959927b54e7189e7227