Disable smart elaboration of negation in guards
Summary:
As title, do not narrow types under negations. The issue is that, currently, `not (V == a)` is refined identically to `(V == a)`, as the negation is not "propagated".

Occurrence typing is still able to soundly refine types under negations, this is thus only a limitation for functions where occ typing is disabled.

Reviewed By: ilya-klyuchnikov

Differential Revision: D58184266

fbshipit-source-id: 4f71dd99971a7ce0dd73e61b0ebda435952fa4a1