Rename erl_scan_otp27 into elp_scan
Summary:
The moment of truth. Take the new scanner, patched, and rename it into `elp_scan`. This causes the new scanner to be in use. The diff should basically be an aggregate of all OTP 27 changes on top of our offset-based scanner.

There may still be issues with the handling of sigils and triple quoted strings, but the most important part here is not to break the existing logic.

Reviewed By: alanz

Differential Revision: D58779869

fbshipit-source-id: 6f0d460650bf806195bc8f97eece43be22a2707c