Upgrade elp_lint to latest available version
Summary:
This was achieved by upgrading the linter to the latest upstream version and by re-applying the following changes.

* [Do not check behaviours](https://www.internalfb.com/phabricator/commit/FBS/d50b6e707a059cbd14260990360d3ed89fb1437a)
* [Avoid crash on empty/new files](https://www.internalfb.com/phabricator/commit/FBS/9cdab82da91ecccd0d219821fc433cc2fd7c8b0c)
* [Do not attempt to lint headers](https://www.internalfb.com/diff/D42917670)

Every other local change to the fork was deemed un-necessary (e.g. support for structured comments was added, then removed)

In addition to that, the module was made OTP-26 compatible by:

* Adding back the explicit enabling for the `maybe` feature
* Stripping away documentation attributes (using triple-quoted strings)
* Reverting [this change](https://github.com/erlang/otp/commit/5343ed1f2043ed2d094458e26c8516a47b94184c) which relied on sigils

The new linter raised a few warnings in WASERVER which were fixed in D57721417.

Reviewed By: alanz

Differential Revision: D57713233

fbshipit-source-id: a03c3b7034da071f82dcf23cf963fabbce39c1b9