add `--one-shot` to elp lint to apply all fixes at once
Summary:
Sometimes we want to put together an ad-hoc diagnostic assist and apply it to all instances in a given file at once.

e.g. If we are changing the order of parameters in a function, where we will keep on getting a diagnostic for it even after making the change.

Provide an "at your own risk" option to apply all the fixes at once.

Due to the way the edits are constructed, it should warn if any changes would overlap.

Reviewed By: jcpetruzza

Differential Revision: D58786579

fbshipit-source-id: 060ce87501c327ad5396d95122cba2c280f9492a