Add DiagnosticConditions.explicit_enable
Summary:
Add a new condition requiring a diagnostic to be explicily enabled before it is active.

This allows us to add diagnostics that may be useful, but then only activate them via the project config.

This diff does not take action on the flag, but it is a noisy one, so it is separated out.

Note: The `DiagnosticConditions` are an explicit list of the conditions a diagnostic is active.  Hence it is kept explicit wherever it is used.

Reviewed By: jcpetruzza

Differential Revision: D59272277

fbshipit-source-id: dc097c68343e955b4b8a7748e4b694bf61d8c20f