Explictly filter explicitly enabled diagnostics
Summary:
If a `DiagnosticDescriptor` has the `explicit_enable` flag set, we do not report it unless its code appears in the `DiagnosticConfig.enabled` field.

This diff does that filtering, and ensures the contents of `DiagnosticsConfig` is updated according to the contents of `'.elp_lint.toml`.

We do not have tests yet,  because no `DiagnosticDescriptor` has this flag set.

They appear later in the stack.

Reviewed By: jcpetruzza

Differential Revision: D59272525

fbshipit-source-id: e47ed143a35a749c4e1ae48e911e8f940d6bec55