Include event start time in telemetry
Summary:
We send telemetry with the duration of various events, and when they occurred.

These show up in SCUBA with a time stamp generated to record the time the message was received in the client.  When working on a remote server, there may be a lag or disconnection which means the time stamp we see is not necessarily the time the event happened in ELP.

Add an additional field to the telemetry to store the start time of the event, as seen in ELP

Reviewed By: robertoaloi

Differential Revision: D59320505

fbshipit-source-id: 619c31165dac70c3732a1a39b05718c200e540e9