Better erlang_service protocol
Summary:
This changes the protocol erlang service uses to communicate between ELP and erlang to be packet and binary based. This will later unlock migrating to a different transport, but for now transport is the same - stdio.

On the ELP side a lot of processing work is moved out of the reader and writer process and into the requesting thread - this should help with scalability. This also significantly simplifies code separating all the concerns and completely removing the concept of a "mismatched sender".

On the Erlang side, segment serialisation work is also moved into the worker thread, out of the main service gen_server - this should also help with scalability.

Reviewed By: alanz

Differential Revision: D58673021

fbshipit-source-id: a0905cd212f12f4d03f7d55ca30293fb1e8921c2