BE: supply range without args to find_cal_in_function callback
Summary:
Often we generate a diagnostic for a function we have found, where the function name is the important thing.  The arguments to the call can be very long and complex.

So facilitate making more focused diagnostics by providing the range of the function without arguments too in the `MakeDiagCtx`

Reviewed By: jcpetruzza

Differential Revision: D59634419

fbshipit-source-id: 1864a6fb247ed06f7015f739a1ca23721d0cf4ac