Fix CallTarget.range for auto-erlang remote call
Summary:
When lowering to HIR, if a local function call is actually to the auto-imported erlang module, we lower it as a remote call with a module of `erlang`.

Because this is a non-existent module, we cannot map back from the `ExprId` to a node in the ast, to get the range.

So adjust `CallTarget::range()` to check for this case, and return the range of the function part only.

Reviewed By: jcpetruzza

Differential Revision: D59637312

fbshipit-source-id: 8861e6afd78c9b78e40a69ed4c8aac713f274fb5