Simplify LSP main loop task processing so `slow_event` telemetry is accurate
Summary:
We currently have telemetry in the main loop if an event takes too long.

We are seeing incidents where we report on a slow task.  But the way the server works, the handler for task events does a private read of the event queue and processes any pending task events.

Remove this private read, so the main loop telemetry is accurate, and the code is simpler.

This should have no functional impact on the code.

Reviewed By: michalmuskala

Differential Revision: D59962692

fbshipit-source-id: 1a109d39fba7341ab356f82f7b2b3168b817399f