erlang service: remove spurious trap_exit call
Summary:
I don't really understand why it's there - this is called in the supervisor, that should be trapping exits already. The process itself didn't do any trapping and was not prepared to handle `EXIT` messages. This seems to be some bad migration.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: alanz

Differential Revision: D60050802

fbshipit-source-id: 3b46b089e966569f6f17d7799d31cbfa5f6571ef