Do not report an error for doc attributes file invalid in erlang module
Summary:
The `erlang` module for OTP 27 has a doc attribute

```
-doc {file,"../../doc/src/erlang_system_info.md"}.
```

The referenced file is not included in all erlang binary packages, so silence the error if it is not found.

Also, some errors emitted by `elp_epp` were referring to `epp`, not `elp_epp`, so were not being formatted correctly.

Reviewed By: RobinMorisset

Differential Revision: D60166779

fbshipit-source-id: 5c2f77bd458396cb826df4c2c6352ac94f636fe8