Remove expr field from structured eqwalizer errors
Summary:
This field is not used on the ELP side. If we need AST of the failing expression we have location - we can use ELP facilities to get the AST.

The field is causing issues since it can be extremely deeply nested for things like list ast (chain of cons cells) which causes panics due to too deep stack.

Reviewed By: ilya-klyuchnikov

Differential Revision: D60519067

fbshipit-source-id: 27703a118787516304e5da9586b2f3750b0eb800