3/n Update request_parse architecture in ELP
Summary:
Modify the ELP `request_parse` infrastructure to run in its own loop, reading and processing callbacks.

We identify a callback by an initial verb (we use "OPN" for this), followed by the original request ID and a data payload.  This is routed to the original ELP `request_parse` via the `inflight` records, but the record is not dropped yet.

Add matching changes in the erlang service.

This is only establishing the channel comms, we do not do anything with the payload on either end yet.

Reviewed By: michalmuskala

Differential Revision: D60450699

fbshipit-source-id: b8814057912d82e3b64de2a3910fd0b665950985