1/n: erlang_service include resolution via ELP
Summary:
This stack is about routing the `file:path_open/3` calls in the erlang service back to ELP for resolution.

The end result is that they are resolved in ELP and the result returned, and then we call the original anyway.  This gives us a simple fallback if for any reason resolution fails in ELP, and the first step of the erlang process is to attempt a straight open, so it achieves the required goal.

The stack results in the following benefits
- Clears the way for not having the erlang service touch the disk at all, when we do the same for `file:open/2`, meaning we can have as-you-type erlang service diagnostics
- Allows more complex `include_lib` processing needed for buck project models to be used in ELP and the erlang service.
  - We have BUCK include targets that construct a mapping of source include files into an output directory that is resolved via normal include lib.  Rather than using the output, which requires a build, we can just resolve the original locations instead.

A gentle start to the stack, adding a callback function in the erlang service only. More to come.

Reviewed By: michalmuskala

Differential Revision: D60450320

fbshipit-source-id: 7756a76941fc5f1629674695d4284758828f715b