12/n Remove erlang_service synthetic include file tracking
Summary:
We had a workaround where the erlang service compile step would tell us what files it had included, and we registered a read on them in salsa to track dependencies.

As we now process the include file resolution via ELP, this happens automatically, to the workaround can be removed.

Reviewed By: michalmuskala

Differential Revision: D60510555

fbshipit-source-id: 278793603d24c70e1ec25a02127df8d4c1707f9d