7&8/n use include file resolution in erlang service
Summary:
Hook up both ends of the reply path
- Use the `IncludeCtx` to resolve the include path in ELP
- crack out the returned path in `elp_epp` and use it to open the file.

Sometimes there is a typo and the include file simply does not resolve.
Deal with this cleanly by falling back in the erlang service to using the original name. It will also fail to resolve, but trigger the normal error reporting process.

Reviewed By: michalmuskala

Differential Revision: D60451832

fbshipit-source-id: 21424ddf259198ab59c2419193e5a302e2155427