erlang_service: remove cursor
Summary:
This removes the need to use `Cursor` and simplifies some other processing around the `Response` type.

When reading a request we now always expect an id + status - this makes implementation much simpler. On the Erlang side this only affects the `EXT` message where we just send some dummy data for those payloads.

Reviewed By: alanz

Differential Revision: D60514366

fbshipit-source-id: 97e82d7770e1280ab03c214dd0ca69116be1918a