Fix transitive check
Summary:
Fix transitive checking of stubs.

It is not sound to assert that a reference is valid if it has already been visited, because the whole cycle may prove to be invalid later on.

The proposed fix is to store "maybe invalid" references (those that have already been visited), and resolve them at once at the end, once information from non-recursive branches has been propagated.

Reviewed By: ilya-klyuchnikov

Differential Revision: D60903935

fbshipit-source-id: 46102b7583f1b98cdc081803cce97dcfe44e17da