Put constructor option into fold parents field
Summary:
When we run a `hir::fold`, we keep track of a list of parents for the current location, to help with assessing context.

Up to now, these were only `AnyExprId`.  But HIR has more strcuture than that, and we sometimes need to have visibility to it.

So track the parents via a `ParentId`, which is now an enum containing either the current `HirIdx`, or a new `ParentId::Constructor`, which will be used later in this stack.

Reviewed By: robertoaloi

Differential Revision: D61591164

fbshipit-source-id: fa37331206fb7e8ed3332eb1e0d8b8310ed65f24