Add diagnostics to avoid 'and' and 'or' operators
Summary:
Unlike other languages, the precedence of `and` and `or` in Erlang is higher than other operators.

There are variants `andalso` and `orelse` which have a more expected precedence.

Generate a warning for the use of `and` or `or`, and offer assists to either ignore the issue or replace the operator with the lower-precedence version.

Note that this also potentially changes the semantics of the programme, as the `andalso`/`orelse` operators have lazy semantics: the RHS is only evaluated if the LHS is insufficient to evaluate the expression.

Reviewed By: robertoaloi

Differential Revision: D61475299

fbshipit-source-id: da7ad489d7bf70949478168bcc9b379adb86f515