Report bad map types
Summary:
Report approximated map types. The error depends on whether the ill-formed association is required or optional.

This is done by adding an invalid form to the module stub during expansion containing the location of the ill-formed association.

Reporting is hidden behind variable `EQWALIZER_REPORT_BAD_MAPS` which defaults to `false`.

Reviewed By: ilya-klyuchnikov

Differential Revision: D62180614

fbshipit-source-id: ad163fa95222efd5f0635da4eb39c3238a82b718