Allow a spec to include a module name, if it matches current module
Summary:
A spec such as

```
-spec modu:fun() -> ok.
```

is allowed, so long as `modu` is the current module.

This diff updates the lowring process to ignore the module name if present. We leave any mismatch checking to the Erlang Service to report, but want to be able to process the code that is there.  The tests are updated to document this.

Reviewed By: TheGeorge, michalmuskala, TD5

Differential Revision: D62231753

fbshipit-source-id: 1d34c3a57fa2b76e1f8b1334ed7804d4b5daace7