6/n Always use ELP file contents in erlang_service
Summary:
This is the final step on the erlang_service side.
It makes `erlang_service_server:path_open/2` functionally equivalent to `file:open/3` with Opts as `[read]`.

It uses a callback to ELP to identify the included file, which returns the file path, its `FileId` and the file text directly.

The nett result is that the erlang service now gets all of its information via IPC, and no longer reads any from on-disk.

Reviewed By: michalmuskala

Differential Revision: D61392274

fbshipit-source-id: 860ba59e33adda446129b3a215af71afef2a70fe