4/n Pass file contents to erlang_service with parse request
Summary:
When requesting the erlang service to parse a file, pass the file contents with the request.  We sent it directly as bytes on the wire, for simplicity and performance.

Unfortunately `elp_epp` processing needs a `pid()` for the open file, so we provide a gen_server to do this.

I suspect the performance can be improved.

Note: It seems that the one glean test is now unexpectedly getting types from the eqwalizer setup, that it did not before.

Reviewed By: michalmuskala

Differential Revision: D60764433

fbshipit-source-id: ab2c1d830387705b167f52bd80af4a8be4bd7046