3/n Add erlang_service callback reply status
Summary:
Use the same status mechanism for callback replies as we do for request replies.
This means that after the original request Id, we have a byte indicating the status, 0 for success, 1 for failure.

Also update the reply message from the e`rlang_service_server` `gen_server`  to use standard `ok` and `error`.

This will be important when we start returning file contents, and must distinguish between a file not found vs a file with empty contents.

Reviewed By: jcpetruzza, michalmuskala

Differential Revision: D60762013

fbshipit-source-id: ee12760b7caf6a870fbd05bd8a7b86e0fd73c883