5/n Make OTP source available in `elp parse-all`
Summary:
We need to OTP project to be available to load files from it in the erlang service.

Do this for `elp parse-all`, and insteady filter out otp files in the main processing loop.

It means the total file count is slightly overstated for the spinner, but no extra work is done.

Reviewed By: michalmuskala

Differential Revision: D60771098

fbshipit-source-id: 37e30267f4e79c5254a5eb8abe1a8f3766a93562