1/n: Tramp FileId into erlang_service compile, for merge
Summary:
# Context

This diff stack is about disconnecting the erlang service from the file system. This is done by passing files to be parsed, and include files to be resolved and opened from ELP to the erlang service via IPC.  This avoids race conditions on the disk, and ensures we always process the "source of truth" in the IDE, as required by the LSP data model.

It potentially opens the way for as-you-type erlang service and eqwalizer diagnostcs, should we choose to enable that.

# This Diff

When resolving include files in ELP we need to know the file we are including from, to be able to work out the correct application, and for relative path includes.  ELP tracks files via `FileId`, which is a `u32` index into the salsa data store.

Since resolution is now happening via a callback from the Erlang Service, make the `FileId` available when the initial parse request is made.

This will allow it to be passed back and used in the callback request, later in this stack.

This change simply makes it available in the `elp_epp` state.

Reviewed By: michalmuskala, robertoaloi

Differential Revision: D62189172

fbshipit-source-id: a0e953341800bdb937327de262684e979deed656