2/n: pass the file_id back to ELP
Summary:
The prior diff sent the ELP `FileId` for a file to the erlang service when the file was opened, and made sure it was in the `elp_lint` state.

We now pass it back to ELP when requesting include file resolution, so we can use the provided one rather than that of the original file in the parse request in `erl_ast`.  This becomes important when we are following a chain of nested includes.

Reviewed By: michalmuskala, robertoaloi

Differential Revision: D62189415

fbshipit-source-id: fd4a83a9160e778a47501e4570231af6098b44a5