BE do not use disk fixture for eqwalizer test
Summary:
When running a test fixture requiring the erlang service, there is no longer any need to put one on disk.

This does that, with the wrinkle that Eqwalizer needs to be able to read the `erlang` module, so we must make sure it is available in the in-memory test fixture. And we must filter it out in the glean tests.

Reviewed By: jcpetruzza

Differential Revision: D62245905

fbshipit-source-id: 19a261fe38cd0edea34eda910e129e1328d5f93f