lint: check for errors after applying fix, bail if they are introduced
Summary:
If the new `--with-check` flag is provided, after applying a fix to the code, check the new diagnostics that we have generated, to ensure they have no errors.

This functionality is protected by a flag as the normal usage is to *fix* errors, so we must allow them to be present.

Reviewed By: robertoaloi

Differential Revision: D62627600

fbshipit-source-id: c067a9e9a5e3c3dac1b67f1f3d97d39faba8c3e1