Do not show inlay hints for "underscored" variables
Summary:
When reporting inlay hints, we currently perform a strict check on the name of the variable, showing them if they do not match.
In Erlang, it is common to prepend variables with an underscore, if they are not used. In that case we should not display the hint if the remaining part of the name matches.

Reviewed By: alanz

Differential Revision: D63695998

fbshipit-source-id: 193cb576fa710a0a89e976bfab69b8cfd1948957