Add regression test for unused-function-args
Summary:
# Context
There was a bug in which function arguments that were used inside only in a pattern were reported as unused.

On investigation, the root cause was a corner-case in the tree-sitter grammar, that would lead to the last element of a list before the |  to be ignored, if it had a match (the tail would get ignored as well).

# This diff
This was fixed in D63701345, here we just add a testcase that would show the issue as regression.

Reviewed By: alanz

Differential Revision: D63701415

fbshipit-source-id: 7db8510b14da9fd9aa33ef93995d954f3aaffd84