Extend check-format linter to custom functions
Summary:
When `io:format/2` is called in a way such that the arguments don't match the template, the Erlang linter (`L1317`) warns the user. But sometimes other functions in the codebase - often invoked as macros - would benefit from the same check.

This extends the forked linter from OTP to take into account extra functions which exhibit the same behaviour.

T203547179 has been filed to allow ELP linters to get the list from a configuration file (or a section in the `.elp.toml`) instead of hard-coding the list.

Reviewed By: alanz

Differential Revision: D63767973

fbshipit-source-id: 6b207be2f7e820c5ae61bd1f40950ccf08dfe366