1/n: Turn fold Strategy into a struct containing an enum
Summary:
This stack is about explicitly representing parens in `hir::Expr`.

This allows is to do analysis for assists without having to switch to the  lower `ast` layer to see parens.

The goal is to treat them in a similar fashion to the existing `hir::Expr::MacroCall`, which is lowered into HIR, but is normally invisible when traversing the HIR ast because of the way we index:
- We implement indexing on `Body` for `ExprId`.  Normally, this will detect that we are entering a `Expr::MacroCall`, and return its expansion.
- If we instead use an `UnexpandedIndex` wrapper around the `Body`, its indexing implementation does not skip over the `MacroCall`.

We will use the same mechanism for `hir::Expr::Paren`.

The first step is to turn the `fold::Strategy` into a struct containing the existing macro expansion strategy only, so we can later extend it to include a strategy for parens.

Reviewed By: jcpetruzza

Differential Revision: D63390669

fbshipit-source-id: 38120d1f2500ec8fcf59032c47f0b6680503f028