4/n: Capture parens for CallTarget::Remote
Summary:
As noted in the previous diff, we have a problem representing parens around a `CallTarget::Remote`, as it has fields for both module and function name, and the parens are around both.

Since capturing parens is for use during analysis of code for assists, primarily for determining if they need to be added, we have two choices
- Do nothing, a remote function call never needs parens, it does not matter
- Capture the fact that there are parens, but not how many levels.

This diff takes the latter approach, to show how it can be done.

Reviewed By: robertoaloi

Differential Revision: D63391932

fbshipit-source-id: e4aa01937bf34fdef588b6816734e76beb6bbd76