3/n: Introduce hir::Expr::Parens
Summary:
This diff extends `hir::Expr` to explicitly capture parens.

It also makes `UnexpandedIndex`, used to control indexing into a `Body` for `ExprId`, take a pair of configuration settings, one for whether macros are visible, and one for whether parens are visible.

These are populated appropriately from the `fold::Strategy` and control what is seen as a fold or indexing operation is done on it.

There is one wrinkle, it is not clear how to deal with lowering a `CallTarget::Remote` which is wrapped in parens as a whole.
The `CallTarget::Local` is fine, as it contains a single `ExprId` so this can be a parened one.  But the remote version has the module and function name captured separately.

A possible solution is provided in the next diff.

And a later diff simplifies the `FoldBody`, as we only need one variant with the explicit config in it.

Reviewed By: robertoaloi

Differential Revision: D63391584

fbshipit-source-id: aed7003ff73226ee82fd72ec68b5f51acd97cf71