2/n: Introduce VisibleParens in Strategy
Summary:
The next step is to introduce a flag for `VisibleParens` into the fold strategy.

This is a mechanical change, it is simply added in all the relevant places with a default value making parens invisible.  It has no effect yet.

The current choice is to make the strategy choices explicit at the point of use, so it is clear when coding what the behaviour will be.

Reviewed By: jcpetruzza

Differential Revision: D63390953

fbshipit-source-id: ec8bf43d35ab1e6e9476b22afbeac3b2226ca163