BE: Split ad hoc diagnostics out from DiagnosticsConfig
Summary:
We have a feature in ELP native diagnostics that we can pass in a `Vec` of `AdHocSemanticDiagnostics`, which is a trait value, i.e. a reference to a function.

These are used only in tests, and are currently stored in `DiagnosticsConfig`.  Because they are a reference, they require a lifetime, which means `DiagnosticsConfig` has to have one too.

In the LSP server, we create a `Snapshot` of the `RootDatabase` and calculate diagnostics using it. If the underlying `RootDatabase` is later updated, all in-flight diagnostics get cancelled, because they will not be fresh.

But this `Snapshot` cannot store anything in it with a lifetime, and we have been re-calculating the `DiagnosticsConfig` from scratch every time we request diagnostics.  That calculation is getting more complex with time, so it is better to do it once only.

To achieve that, we split the `AdHocSemanticDiagnostics` out of the `DiagnosticsConfig` structure.

The next diff will then put it into the `Snapshot`

Reviewed By: robertoaloi

Differential Revision: D63833398

fbshipit-source-id: 0f4e93c69a1c6d215ba1345cb4f433f2b7be4475