BE: Put DiagnosticsConfig into server Snapshot
Summary:
Now that `DiagnosticsConfig` does not have a lifetime on it, we can include it in the `Snapshot`, rather than calculating a fresh copy every time we ask for anything needing it.

This also opens the path for putting richer content in `DiagnosticsConfig`, and subsequent diffs will add diagnostic severity remapping info.

Reviewed By: robertoaloi

Differential Revision: D63833770

fbshipit-source-id: a6475153d861bf02e8a88f3f11a38e9038d71607