Filter stub forms from AST instead of relying on the STUB section
Summary:
In the past, it was not possible to decode/filter the AST on the Rust size in an easy way. Therefore, to optimize the communication between EqWAlizer and ELP we were producing a `ParseResult` with two distinct entries: the full AST and a _Stub_ (a sub-set of the AST containing only attributes interesting for Eqwalizer, such as type definitions and specs). This is no longer necessary as a filtering mechanism is already in place in the Rust side of eqWAlizer.

This diff switches to use a filtered version of the AST instead of the Stub, which will be cleaned up in a separate diff in the stack.

Reviewed By: VLanvin

Differential Revision: D64110578

fbshipit-source-id: 75660c55905d7d8703c28696ad62875bae786231