Introduce record tuple mismatch diagnostic
Summary:
A record is just syntax sugar over a tuple.

So you can match a record either using record syntax, or as a tuple.

Matching as a tuple is brittle though, if the record changes, the match may silently fail.

Generate a diagnostic when a record is matched as a tuple.

Reviewed By: robertoaloi

Differential Revision: D64602534

fbshipit-source-id: 2187001a685be2a62ccfa86494f353f8313f7a77