Switch ELP from build_info to manifest plugin
Summary:
Now that `rebar3` has been bumped to `3.24.0`, we can use the built-in manifest plugin instead of the `build_info` one.

Given the limited user base for ELP, I think it's safe to have this breaking change and require users to bump their rebar3 version, instead of supporting both the native and additional plugin.

Reviewed By: michalmuskala

Differential Revision: D60914619

fbshipit-source-id: 08902fa93125329919d1a3734b171fe48e77b743