BE: rename API call for hir::Var
Summary:
Sometimes in the course of writing an assist we need to rename a variable. The existing rename infrastructure works from a `SymbolDefinition`, which is not naturally available in HIR.

Create `rename_rename_var()` as a simple way to do the rename, based on HIR data sources only.

Note: the dead code marker will be removed in a later diff, when it is used for an assist

Reviewed By: jcpetruzza

Differential Revision: D64825381

fbshipit-source-id: 585b21b00ea059ed78b70986fad23a5162a72449