BE: improve internal rename API
Summary:
Our internal rename API had a callback

```
get_name: &dyn Fn(Option<&ast::Name>) -> String,
```

This is needed for the specific case of inlining an expression, where we replace a variable with the corresponding expression, and need to check at each replacement location if parens are needed.

Make the situation clearer by always passing in the new name, and adding a callback that asks if parens are needed at the specific location instead.

Reviewed By: jcpetruzza

Differential Revision: D64829976

fbshipit-source-id: 7d22279f39cb01823e595f27abf6ad0c036f6003