Check minimum rebar3 version
Summary:
Instead of checking for the existence of the experimental manifest provider, verify the current rebar3 version matches the minimum required version.
Also split the version extraction into its own function.
Use the `semver` crate for the comparison. Finally, ensure meaningful messages are returned to the user in case of issues extracting the version.

Reviewed By: michalmuskala

Differential Revision: D64842181

fbshipit-source-id: 2abddec7cb824415c8b09f90590fdb2819f255af