2/2 Remove redundant wid:make/1 call
Summary:
When changing a function parameter from `wa:who()` to `wa:wid()`, remove any cal to `wid:make(WHO)` called for that paremeter.

Only do so if it is a simple match, and the parameter is not used anywhere else.

If those conditions hold, remove it, and rename the parameter to the LHS of the original match.

Reviewed By: robertoaloi

Differential Revision: D64908029

fbshipit-source-id: fa1cc7f47dae9ee920023db26bf5aea22f094d43