BE: fold: introduce ParentId::TopLevel
Summary:
When the fold callback is invoked, the context has a list of parent nodes up to the root of the tree.

It is useful to ask for the current parent, as a function, and not have it return an `Option` type.

So introduce `ParentId::TopLevel` to ensure the list is always non-empty, and then have `ctx.parent()` return a `ParentId` without failing.

Reviewed By: robertoaloi

Differential Revision: D64905551

fbshipit-source-id: a7bf596df2449414e12cb949b173e890b7433ccb