Add IncludeFileIndex for resolving includes from paths
Summary:
We create a cache in salsa of the mapping from full .hrl file path to its corresponding `FileId`, and use this in include file resolution.

Like the `ModuleIndex`, we only build this when the project structure changes, or the contents of a given `SourceRoot` changes.

This will facilitate later changes when we improve the Buck project model.

Interestingly, it improves our test fixture fidelity, causing a test to be updated because otherwise it now reports an unused include file error.

Reviewed By: jcpetruzza

Differential Revision: D64532901

fbshipit-source-id: 8c37485abfb7698949a11625b85f5fe9fa73a5aa