6/n semantic token types: Bring in 'type_dynamic' semantic token tag
Summary:
We now that we have cheap, reliable semantic token highlighting of the Eqwalizer `dynamic/0` and `dynamic/1` types, but re-using the semantic tag intended for bound variables in patterns.

We introduce a new tag, `type_dynamic`, and work it through into both the internal and OSS VS Code extensions to highlight in bold.

This is chosen as it does not clash with the bound variables which uses underline, and it is theme-independent.

 https://pxl.cl/5R4Nj

Reviewed By: robertoaloi

Differential Revision: D65215906

fbshipit-source-id: cf80de49a6d15e70c98250bdb12351d94884eb7a