4/n semantic token types: Use eqwalizer types for highlighting
Summary:
Rather than explicitly requesting the eqwalizer type information for a file every time semantic tokens are requested, which is a huge computational burden, use the types we get when we eqwalize the file in the normal course of business.

These were cached in the server snapshot in the prior diff.

Reviewed By: robertoaloi

Differential Revision: D65215962

fbshipit-source-id: 04424cabb1e33a1ff36525022549850e11fdef96