5/n semantic token types: refresh semantic tokens when they change
Summary:
Our Eqwalizer type caching approach has a race condition problem with the semantic tokens request.

The client requests these when the contents of the file changes (or the visible part changes if the viewport is scrolled).

We only calculate types when the file is saved, and it is some indeterminate point that we get the results.

Luckily the LSP spec comes to the rescue, and we can send a server request to the client (VS Code) to ask it to request fresh semantic tokens.

This now brings in unobtrusive semantic highlighting of the eqwalizer dynamic types.

It still uses an incorrect semantic tokenn tag, this will be remedied in the next diff.

Reviewed By: robertoaloi

Differential Revision: D65216374

fbshipit-source-id: 5d3a349ebec5fbe5c86bc9835720d68a0e563d95