BE: Introduce IncludeGenerated enum instead of bool for eqwalizer calls
Summary:
We currently use a bool to control whether Eqwalizer should process generated files or not.

Replace this with an enum instead, so it is clear what is being requested, and where the requests originate.

This also makes it easier in future to extend the enum if we need additional control points.

Reviewed By: jcpetruzza

Differential Revision: D65336558

fbshipit-source-id: 28f3c622503c60000f84b60c9143b469dc7518bd