BE: Switch to using a struct for eqwalizer include generated
Summary:
We have been passing a variable called `include_generated` to the various eqwalizer functions to control whether it should eqwalize generated files too.

In an earlier diff we made this variable an enum called `IncludeGenerated`.

This diff replaces it with a struct, having a single field containing the original `include_generated` value.

This positions us to add additional criteria in the next diff

Reviewed By: robertoaloi

Differential Revision: D65342308

fbshipit-source-id: 8ed7728521e61e7d8676bd83abd641f06c5848ca