Add support for EEP 059 documentation (OTP 27)
Summary:
If the current Erlang/OTP runtime supports [EEP 059](https://www.erlang.org/eeps/eep-0059) doc attributes (i.e. the user is running OTP 27 or higher), try to extract them as part of the documentation request.

To do so, we pass an extra argument to the `DocRequest`, which is the `AST` originally returned by a `ParseRequest`. This allows us to compute documentation leveraging the internal `beam_doc` module, without recomputing the AST.

We take this opportunity to cleanup the code in the `erlang_service_lint` module, removing unused variables and splitting the existing functionality in multiple functions.

Notice that, if at least one EEP 059 attribute is used in a module, old style documentation is not rendered (i.e. we don't support a combination of both styles within the same module). We could extract both styles and merge them, but it feels like a wasted effort.

The doc warnings are still unhandled at this stage.

This should only be considered as a temporary solution, since in T206726412 we plan to extract the documentation attributes natively in ELP.

Reviewed By: michalmuskala

Differential Revision: D64242690

fbshipit-source-id: 1f87e41c6bb4869e6b7240ef075182563a7e3ab2